pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Василий Серебрянников <svaaugust@gmail.com>"]
readme = "README.md"
packages = [
    { include = "shared" },
    { include = "tests" },
]
[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.3"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.5.2"
httpx = {extras = ["socks"], version = "^0.27.2"}
aiosqlite = "^0.20.0"
jwt = "^1.3.1"
pydantic = {extras = ["email"], version = "^2.9.2"}
pytest-asyncio = "^0.24.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.12"
commitizen = "^3.29.1"

[tool.poetry.scripts]
alembic = "alembic.config:main"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
